requirements.txt
# Framework Principal
discord.py>=2.4.0

# Gerenciamento de Variáveis de Ambiente (Segurança)
python-dotenv==1.0.0

# Áudio e Extração (Música)
yt-dlp==2025.01.26
PyNaCl==1.5.0

# Utilitários (Opcional, mas recomendado para performance)
aiohttp>=3.9.1